Be more c++11 compatible.
This is what 9.4.2/2 says:

"The declaration of a static data member in its class definition is not a
definition and may be of an incomplete type other than cv-qualified void. The
definition for a static data member shall appear in a namespace scope enclosing
the member’s class definition"

So, using magical IoC macros for config outside of pd::config namespace is
actually prohibited by c++11.